Refactor to use modules for instance and class methods, bump version to 1.3.0
This changes ThirdBase::Date and ThirdBase::DateTime to put all
of their class and instance methods in modules, in order to make
them easier to override while allowing the use of super.  It also
makes including the compat part affect ThirdBase::Date and
ThirdBase::DateTime, and makes ::Date and ::DateTime just a simple
subclass of those two classes.